convert all widgets to typescript
