+++ lektor/admin/static/js/widgets/mixins.tsx
@@ -8,7 +8,15 @@ export const widgetPropTypes = {
   disabled: PropTypes.bool,
 };
 
-export function getInputClass(type) {
+export type WidgetProps<ValueType = string> = {
+  value?: ValueType;
+  type?: any;
+  placeholder?: ValueType;
+  onChange: (value: ValueType) => void;
+  disabled?: boolean;
+};
+
+export function getInputClass(type: { size: string }) {
   let rv = "form-control";
   if (type.size === "small") {
     rv = "input-sm " + rv;
